Whisper: rename whisper to mlx_whisper
