+++ whisper/mlx_whisper/assets/download_alice.sh
