+++ whisper/mlx_whisper/timing.py
